Solution to Logical NOT Operation

+++ Logical NOT Operation@@67882a731b4392d8fc1577de/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main() {
-    float a;
-    scanf("%f",&a);
+    int a;
+    scanf("%d",&a);
     if (a>!0) {
         printf("True");
     }
